improved AboutWidget a little bit

+++ retroshare-gui/src/gui/AboutWidget.h
@@ -114,7 +114,7 @@ private:
     bool			mImagesReady ;
     int				mState;
     int 			mTimerId;
-    int 			mStep;
+    float			mStep;
     int 			mMaxStep;
 };
